Main fixes for Django 2.2 on python3.
This isn't a complete fix as one of the unittests is failing
due to a change of order of iteration around children on Py2/Dj2.2
although is still OK on Py2 / Dj1.11
