+++ django_dag/models.py
@@ -286,8 +286,8 @@ def edge_factory(node_model, child_to_field = "id", parent_to_field = "id", conc
         class Meta:
             abstract = not concrete
 
-        parent = models.ForeignKey(node_model, related_name = "%s_child" % node_model_name, to_field = parent_to_field)
-        child = models.ForeignKey(node_model, related_name = "%s_parent" % node_model_name, to_field = child_to_field)
+        parent = models.ForeignKey(node_model, related_name = "%s_child" % node_model_name, to_field = parent_to_field, on_delete=models.CASCADE)
+        child = models.ForeignKey(node_model, related_name = "%s_parent" % node_model_name, to_field = child_to_field, on_delete=models.CASCADE)
 
         def __unicode__(self):
             return u"%s is child of %s" % (self.child, self.parent)
